add condition tests, update verifier for Age data type
