+++ argonaut-api/src/test/java/gov/va/api/health/argonaut/api/resources/ConditionTest.java
@@ -1,0 +1,63 @@
+package gov.va.api.health.argonaut.api.resources;
+
+import static gov.va.api.health.argonaut.api.RoundTrip.assertRoundTrip;
+
+import gov.va.api.health.argonaut.api.ZeroOrOneVerifier;
+import gov.va.api.health.argonaut.api.bundle.AbstractBundle;
+import gov.va.api.health.argonaut.api.bundle.BundleLink;
+import gov.va.api.health.argonaut.api.resources.Condition.Bundle;
+import gov.va.api.health.argonaut.api.samples.SampleConditions;
+import java.util.Collections;
+import lombok.extern.slf4j.Slf4j;
+import org.junit.Test;
+
+@Slf4j
+public class ConditionTest {
+
+  private final SampleConditions data = SampleConditions.get();
+
+  @Test
+  public void bundlerCanBuildConditionBundles() {
+    Condition.Entry entry =
+        Condition.Entry.builder()
+            .extension(Collections.singletonList(data.extension()))
+            .fullUrl("http://condition.com")
+            .id("123")
+            .link(
+                Collections.singletonList(
+                    BundleLink.builder()
+                        .relation(BundleLink.LinkRelation.self)
+                        .url(("http://condition/1"))
+                        .build()))
+            .resource(data.condition())
+            .search(data.search())
+            .request(data.request())
+            .response(data.response())
+            .build();
+
+    Bundle bundle =
+        Condition.Bundle.builder()
+            .entry(Collections.singletonList(entry))
+            .link(
+                Collections.singletonList(
+                    BundleLink.builder()
+                        .relation(BundleLink.LinkRelation.self)
+                        .url(("http://condition.com/2"))
+                        .build()))
+            .type(AbstractBundle.BundleType.searchset)
+            .build();
+
+    assertRoundTrip(bundle);
+  }
+
+  @Test
+  public void condition() {
+    assertRoundTrip(data.condition());
+  }
+
+  @Test
+  public void relatedGroups() {
+    ZeroOrOneVerifier.builder().sample(data.condition()).fieldPrefix("onset").build().verify();
+    ZeroOrOneVerifier.builder().sample(data.condition()).fieldPrefix("abatement").build().verify();
+  }
+}
